requirements.txt
aioboto3
aiobotocore
aiofiles
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alembic
annotated-types
anyio
async-timeout
asyncpg
attrs
boto3
botocore
certifi
click
databases
exceptiongroup
fastapi
frozenlist
greenlet
h11
httpcore
httpx
idna
jmespath
Mako
MarkupSafe
multidict
psycopg2
psycopg2-binary
pydantic
pydantic-settings
pydantic_core
python-dateutil
python-dotenv
python-magic
python-multipart
s3transfer
six
sniffio
SQLAlchemy
starlette
typing_extensions
urllib3
uvicorn
wrapt
yarl
